+++ eng/scripts/CodeChecks.ps1
@@ -100,8 +100,8 @@ try {
             $status = $status -replace "`n","`n    "
             LogError "Generated code is not up to date.`n" + `
                 "You may need to rebase on the latest master, `n" + `
-                "run 'eng\scripts\Update-Snippets.ps1' if you modified sample snippets or other *.md files, `n" + `
-                "run 'eng\scripts\Export-API.ps1' if you changed public APIs `n" +
+                "run 'eng\scripts\Update-Snippets.ps1' if you modified sample snippets or other *.md files (https://github.com/Azure/azure-sdk-for-net/blob/master/CONTRIBUTING.md#updating-sample-snippets), `n" + `
+                "run 'eng\scripts\Export-API.ps1' if you changed public APIs (https://github.com/Azure/azure-sdk-for-net/blob/master/CONTRIBUTING.md#public-api-additions) `n" +
                 "run 'dotnet build /t:GenerateCode' to update the generated code."
         }
     }